Rewrite first part of DEVELOPERS.md to use mise
Note: We should delete `.envrc` after a transition period. If `asdf`
is uninstalled but `direnv` remains active, it will complain that `use
asdf` doesn't work. This emits an annoying message, but has no other
harmful effects.

Motivation for this rewrite:

`asdf-direnv` has not been updated for the Go-rewritten version of
`asdf`. This creates problems for anyone following these
instructions. I originally used `asdf-direnv` because some programs
(if memory serves, the `pq` gem compilation running `pg_config`) don't
work when shimmed. I also preferred the speed of `$PATH` manipulation
over using shim programs.

The author of `asdf-direnv` stopped maintaining it months ago, since
`mise` offers a similar model to `asdf` but is based on `$PATH` rather
than shims. As a result, he switched to `mise`, and now so are we.

There are generally fewer steps when using `mise`, but I made
`DEVELOPERS.md` a bit longer to include shell activation instructions
and system dependencies, so the instructions have grown slightly
overall. Hopefully, this will speed up the process, since it is less
necessary to visit other manual pages.

Another major difference is that `mise install` tries to download and
install *every* dependency, so it’s necessary to set up system
dependencies first. This required some reordering and restructuring.

Finally, while it was theoretically possible to manage Postgres by
other means, having it in `.tool-versions` made this awkward and often
painful in practice—even though the previous manual implied it should
be easy. This has now been changed to use the `mise.local.toml`
feature as an opt-in.