Use a thread pool + shared queue for monitor
Because the pulse threads for PostgresServer and MinioServer instances
were counted by the thread calculation, it appears as though if the
number of PostgresServer and MinioServer instances to monitor exceeded
Config.max_monitor_threads - 2, that the monitor would sleep
indefinitely previously (at least, until one of the pulse threads
exited). This change eliminates that possibility.  However, the
tradeoff is the number of threads in the process can exceed
Config.max_monitor_threads.  Config.max_monitor_threads now is only
used to set the size of the monitor thread pool.

To prevent arbitrarily large memory usage if the thread pool is not
able to process jobs fast enough, the queue size is limited to
the thread pool size plus 1.5 times the number of monitored objects.
Once that limit is hit, pushing to the queue will block.  However,
if this happens, this will lead to a greater than 5 second delay
between pulse checks.

Reduce scope of mutex.synchronize calls in the main monitor loop.