Allow Prog::Vm::Nexus.assemble to choose cloud-hypervisor version
Additionally, allow it to choose whether to support hugepages and to
choose the firmware version (though currently only one firmware
version is supported).  The default for hugepages is still true in
production/test and false in development.  Add setting for hugepages,
ch_version, and firmware_version to the strand frame, and pass the
options in the setup-vm prep command.

In rhizome prep_host, download all supported cloud-hypervisor
versions and firmware versions.  Currently, only one firmware
version is supported (no change in supported firmware versions in
this commit, only infrastructure to allow for supporting multiple
firmware versions).  On Ubuntu 24, both cloud-hypervisor 35.1
and 45.0 are installed, with 45.0 as the default if it is installed.
On Ubunutu 22, only cloud-hypervisor 35.1 is installed, and it
is the default.

For CloudHypervisor::{Firmware,Version}, remove the DEFAULT constant,
and add a POTENTIAL constant.  POTENTIAL lists the versions that
should be installed.  Instead of a DEFAULT constant, the default
value of the SUPPORTED hash now has the default version.  A version
will only be selected as the default if it has been downloaded. Add
{Firmware,Version}.[] methods for returning the appropriate instance,
which vm_setup uses to select the appropriate versions.  Add
{Firmware,Version}.download methods for prep_host to use to download
all supported versions.

Fix cloud-hypervisor 35.1 disk setup, by joining params with spaces
instead of newlines.

Use a more robust cloud-hypervisor version check using Gem::Version,
which will work after cloud-hypervisor version reaches 100.

Add CloudHypervisor::Version#{lsb_release,ubunutu_version} to allow
for easier testing/mocking.

To allow Prog::Vm::UpdateIpv6#rewrite_persisted to work, extract
the options handling code to Prog::Vm::Nexus.setup_vm_options_str.
Not sure it's a great idea to look at a VM strand's stack in another
prog, but the alternative would be to add this information to Vm
itself.