Make setup-vm read hugepages/ch_version/firmware_version from params json
This removes the awkward Prog::Vm::Nexus.setup_vm_options_str
class method, as well as reaching into the vm strand's frame to
get the data to provide it in Prog::Vm::UpdateIpv6#rewrite_persisted.

For setup-vm delete* actions, allow the params json file to not exist,
because those actions did not require it exist previously, as parsing
of the params was done after dispatching to those methods.  It now
must be done first, because VmSetup#purge_without_network needs to
know whether hugepages were used.