Remove redundant label in nic_nexus
vm_allocated semaphore and its label wait_allocation is removed from
nic_nexus.

Based on the interactions between vm, subnet and nic nexuses, it was
serving no purpose and the other semaphore in nic_nexus (start_rekey)
is enough to make sure the vm is allocated.

On very rare conditions, this extra step was causing issues in subnets:
If we had two nic strands and one of them was unlucky enough to not get
CPU time and reach the wait_setup label, it would miss the subnet
nic_to_rekey funcion (filtering for nic strands on wait_setup or wait
label) and since decrementing a semaphores, zeroes it out, it wouldn't be
included in the nics-to-rekey array.

Nic logic is shortened and now simpler.